Added edit and delete buttons on climb-log-details.html page.
Created two new buttons--edit and delete--on the climb-log-details.html page. The edit button opens a new page to allow the user to edit their log; the delete button actives a JS script, opening a confirmation popup to double-check if the user truly wants to delete the entry. Upon a "yes" it executes the correct url for Django to run the function to delete the log.